Move geotool classes to Java side, allowing resource loading.
